+++ .formatter.exs
@@ -1,0 +1,5 @@
+# Used by "mix format"
+[
+  locals_without_parens: [project: 2],
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
+]