+++ lib/commanded/scheduler/schedule/schedule_cancelled.ex
@@ -2,10 +2,10 @@ defmodule Commanded.Scheduler.ScheduleCancelled do
   @moduledoc false
 
   @type t :: %__MODULE__{
-    schedule_uuid: String.t(),
-    name: String.t()
-  }
-  @derive [Poison.Encoder]
+          schedule_uuid: String.t(),
+          name: String.t()
+        }
+  @derive Jason.Encoder
   defstruct [
     :schedule_uuid,
     :name