+++ lib/commanded/scheduler/schedule/scheduled_once.ex
@@ -1,5 +1,6 @@
 defmodule Commanded.Scheduler.ScheduledOnce do
   @moduledoc false
+  alias Commanded.Scheduler.{Convert, ScheduledOnce}
 
   @type t :: %__MODULE__{
           schedule_uuid: String.t(),
@@ -8,7 +9,7 @@ defmodule Commanded.Scheduler.ScheduledOnce do
           command_type: String.t(),
           due_at: NaiveDateTime.t()
         }
-  @derive [Poison.Encoder]
+  @derive Jason.Encoder
   defstruct [
     :schedule_uuid,
     :name,
@@ -16,18 +17,16 @@ defmodule Commanded.Scheduler.ScheduledOnce do
     :command_type,
     :due_at
   ]
-end
 
-alias Commanded.Scheduler.{Convert, ScheduledOnce}
+  defimpl Commanded.Serialization.JsonDecoder do
+    def decode(%ScheduledOnce{} = once) do
+      %ScheduledOnce{command: command, command_type: command_type, due_at: due_at} = once
 
-defimpl Poison.Decoder, for: ScheduledOnce do
-  def decode(%ScheduledOnce{} = once, _options) do
-    %ScheduledOnce{command: command, command_type: command_type, due_at: due_at} = once
-
-    %ScheduledOnce{
-      once
-      | command: Convert.to_struct(command_type, command),
-        due_at: NaiveDateTime.from_iso8601!(due_at)
-    }
+      %ScheduledOnce{
+        once
+        | command: Convert.to_struct(command_type, command),
+          due_at: NaiveDateTime.from_iso8601!(due_at)
+      }
+    end
   end
 end