+++ test/support/example_command.ex
@@ -1,8 +1,9 @@
 defmodule Commanded.Scheduler.ExampleCommand do
   @moduledoc false
-  
+
+  @derive Jason.Encoder
   defstruct [
     :aggregate_uuid,
-    :data,
+    :data
   ]
 end