+++ test/support/example_domain/ticketing/timeout_reservation_handler.ex
@@ -11,18 +11,15 @@ defmodule ExampleDomain.TimeoutReservationHandler do
   @doc """
   Timeout the ticket reservation after the expiry date/time.
   """
-  def handle(
-    %TicketReserved{ticket_uuid: ticket_uuid, expires_at: expires_at},
-    _metadata)
-  do
-    timeout_reservation = %TimeoutReservation{
-      ticket_uuid: ticket_uuid
-    }
+  def handle(%TicketReserved{} = event, _metadata) do
+    %TicketReserved{ticket_uuid: ticket_uuid, expires_at: expires_at} = event
+
+    timeout_reservation = %TimeoutReservation{ticket_uuid: ticket_uuid}
 
     schedule_once = %ScheduleOnce{
       schedule_uuid: "schedule-" <> ticket_uuid,
       command: timeout_reservation,
-      due_at: expires_at,
+      due_at: expires_at
     }
 
     AppRouter.dispatch(schedule_once)