+++ test/support/factory.ex
@@ -11,7 +11,7 @@ defmodule Commanded.Scheduler.Factory do
     schedule_uuid = UUID.uuid4()
     schedule_name = "timeout_reservation"
     ticket_uuid = Map.get(context, :ticket_uuid, UUID.uuid4())
-    due_at = NaiveDateTime.utc_now() |> add_minutes(10)
+    due_at = NaiveDateTime.utc_now() |> add_minutes(10) |> NaiveDateTime.truncate(:second)
 
     command = %TimeoutReservation{
       ticket_uuid: ticket_uuid